requirements.txt
PyMySQL==0.10.1
allure_python_commons==2.8.16
requests==2.24.0
jsonpath==0.82
pytest==5.4.3
PyYAML==5.3.1
